stirrer added
